SelectBar
기존에 두개로 나뉘었던, 선택 창을
하나로 통합했습니다.
