+++ src/components/AddHistory.js
@@ -263,7 +263,6 @@ const AddHistory = () => {
 
     return(
         <>   
-            {console.log(store.getState())}
             <Wrapper>
                     <BtnWrapper>
                         <IncomeBtn active={type} onClick={onClickIncome}>Income</IncomeBtn>
